Adds support for wasm32 target
Since wasm32 is current single-threaded in the browser, futures do not
require (and do not implement) the Send and Sync traits. This applies to
reqwest. For that reason, it is necessary to add some conditionals to
detect wasm32 building and disable any Send and Sync type constraints to
get things to work.