ci: use `cargo hack` feature powerset in CI
waiting on #175 & #176 to fix issues with the build

Recently a [change was introduced that did not compile](https://github.com/TrueLayer/reqwest-middleware/pull/167/files/4a2b3f335e9ea2f080f0d13ee782b830607b1f60#r1689595502) #167.

This was not caught by our CI becuase the feature was not enabled when building in CI.

This change ensures that every possible feature combination is attempted when running tests and clippy in CI. This adds up to 50 attempts for all of our features accross the three crates in this repo. This doesn't take too long since they share build cache.